+++ service/controller/rbac/rbac.go
@@ -1,19 +1,14 @@
 package rbac
 
 import (
-	"fmt"
-
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
-	"github.com/giantswarm/k8smetadata/pkg/label"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"github.com/giantswarm/operatorkit/v7/pkg/controller"
 	"github.com/giantswarm/operatorkit/v7/pkg/resource"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/labels"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	legacylabel "github.com/giantswarm/rbac-operator/pkg/label"
 	"github.com/giantswarm/rbac-operator/pkg/project"
 )
 
@@ -43,11 +38,6 @@ func NewRBAC(config RBACConfig) (*RBAC, error) {
 
 	var namespaceAuthController *controller.Controller
 	{
-		selector, err := labels.Parse(fmt.Sprintf("%s,%s", label.Organization, legacylabel.LegacyCustomer))
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-
 		c := controller.Config{
 			K8sClient: config.K8sClient,
 			Logger:    config.Logger,
@@ -55,7 +45,6 @@ func NewRBAC(config RBACConfig) (*RBAC, error) {
 				return new(corev1.Namespace)
 			},
 			Resources: resources,
-			Selector:  selector,
 
 			// Name is used to compute finalizer names. This here results in something
 			// like operatorkit.giantswarm.io/rbac-operator-rbac-controller.