+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bump `giantswarm/operatorkit` to `v7.0.1`.
 - Bump k8s dependencies to `v0.20.15`.
 - Bump `controller-runtime` to `v0.8.3`.
+- Selector moved from watcher to individual handlers for rbac controller.
 
 ## [0.23.0] - 2022-03-02