+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.24.0] - 2022-03-30
+
 ### Added
 
 - Add `externalresources` resource that binds `organization-$org-read` role for any subject with org-namespace access.
@@ -290,7 +292,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Read-only role for customer access into Control Plane.
 
-[Unreleased]: https://github.com/giantswarm/rbac-operator/compare/v0.23.0...HEAD
+[Unreleased]: https://github.com/giantswarm/rbac-operator/compare/v0.24.0...HEAD
+[0.24.0]: https://github.com/giantswarm/rbac-operator/compare/v0.23.0...v0.24.0
 [0.23.0]: https://github.com/giantswarm/rbac-operator/compare/v0.22.1...v0.23.0
 [0.22.1]: https://github.com/giantswarm/rbac-operator/compare/v0.22.0...v0.22.1
 [0.22.0]: https://github.com/giantswarm/rbac-operator/compare/v0.21.0...v0.22.0